+++ microlens/CHANGELOG.md
@@ -1,3 +1,7 @@
+# unreleased (0.3.0.0)
+
+* Moved `Lens.Micro.Classes` into `Lens.Micro.Internal`.
+
 # 0.2.0.0
 
 * Removed `toListOf`.