+++ microlens/CHANGELOG.md
@@ -1,6 +1,7 @@
 # unreleased (0.3.0.0)
 
 * Moved `Lens.Micro.Classes` into `Lens.Micro.Internal`.
+* Added `<%~`, `<<%~`, `<<.~`.
 
 # 0.2.0.0